Rename folders in uppercases
