+++ VIEW/addpost.php
@@ -1,6 +1,6 @@
 <?php require "header.php"; ?>
 <?php
-require "../model/admin/db.php";
+require "../MODEL/admin/db.php";
 
 $titreErreur = $contenuErreur = $auteurErreur = $photoErreur = $titre = $contenu = $auteur = $photo = "";
