+++ VIEW/modal_contact_form.php
@@ -1,6 +1,6 @@
 <!-- MODAL FOR CONTACT -->
 <!-- Modal -->
-<?php require "../controller/contact.php";?>
+<?php require "../CONTROLLER/contact.php";?>
 <div class="modal fade" id="myModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
     <div class="modal-dialog">
         <div class="modal-content">
